# swaggerOut.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Wahlhalla: Task Tracking API",
    "version": "0.0.1"
  },
  "swagger": "2.0",
  "paths": {
...
    "/api/person": {
      "get": {
        "tags": [
          "Person"
        ],
        "summary": "returns a list of all persons",
        "description": "Retrieves a list of all persons.",
        "responses": {
          "200": {
            "description": "a list of all persons ordered by last name.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Person"
                  }
                }
              }
            }
          },
          "400": {
            "description": "validation error"
          },
          "500": {
            "description": "unknown error"
          }
        }
      }
    },
    "/api/person/{id}": {
      "get": {
        "tags": [
          "Person"
        ],
        "summary": "returns a person by ID.",
        "description": "Retrieves person by ID.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Numeric ID of the person to retrieve.",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "a person found by specified ID.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Person"
                }
              }
            }
          },
          "400": {
            "description": "validation error"
          },
          "404": {
            "description": "Cannot find person with id={id}."
          },
          "500": {
            "description": "unknown error"
          }
        }
      },
      "put": {
        "tags": [
          "Person"
        ],
        "summary": "updates a person by ID.",
        "description": "Updates person by ID.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Numeric ID of the person to update.",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "description": "Updated Person",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Person"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Person was updated successfully."
          },
          "400": {
            "description": "validation error"
          },
          "404": {
            "description": "Cannot find person with id={id}."
          },
          "500": {
            "description": "unknown error"
          }
        }
      },
      "delete": {
        "tags": [
          "Person"
        ],
        "summary": "deletes an person by the provided ID.",
        "description": "Deletes an person by the provided ID.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Numeric ID of the person to delete.",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Person was deleted successfully!"
          },
          "400": {
            "description": "validation error"
          },
          "404": {
            "description": "Cannot delete Person with id={id}. Maybe Person was not found!"
          },
          "500": {
            "description": "Could not delete person with id={id}"
          }
        }
      }
    },
    "/api/person/": {
      "post": {
        "tags": [
          "Person"
        ],
        "summary": "creates an person with the provided parameters.",
        "description": "Creates an person with the provided parameters",
        "requestBody": {
          "description": "Person To Create",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Person"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Returns the newly created person.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Person"
                }
              }
            }
          },
          "400": {
            "description": "validation error"
          },
          "500": {
            "description": "unknown error"
          }
        }
      }
    },
    "/api/person/name/{name}": {
      "get": {
        "tags": [
          "Person"
        ],
        "summary": "returns a person by name.",
        "description": "Retrieves person by name.",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "description": "Name of the person to retrieve.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "a person found by specified name.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Person"
                  }
                }
              }
            }
          },
          "400": {
            "description": "validation error"
          },
          "500": {
            "description": "unknown error"
          }
        }
      }
    },
...
  },
...
  "securityDefinitions": {},
...
}